+++ sdk/deviceregistry/iot-device-registry-software-update/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- Initial beta release of the Azure Software Update for Device Registry client library for JavaScript. [#39647](https://github.com/Azure/azure-sdk-for-js/pull/39647)
+
 ### Breaking Changes
 
 ### Bugs Fixed